examples/4_publishing/my_library/pyproject.toml
# THESE ARE JUST COPIED FROM DATABASE LIBRARY, ADAPT THEM
[build-system]
requires = [
    "setuptools>=61.0",
    "robotframework"
    ]
build-backend = "setuptools.build_meta"

[project]
name = "robotframework-databaselibrary"
authors = [{name="Franz Allan Valencia See", email="franz.see@gmail.com"},
]
description = "Database Library for Robot Framework"
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "robotframework",
    "robotframework-excellib"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
license = {text = "Apache License 2.0"}
dynamic = ["version"]

[project.urls]
"Homepage" = "https://github.com/MarketSquare/Robotframework-Database-Library"
"Keyword docs" = "http://marketsquare.github.io/Robotframework-Database-Library/"

[tool.setuptools.dynamic]
version = {attr = "DatabaseLibrary.__version__"}

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
